Rename module ykman.cli.logging_setup to ykman.logging_setup
